Fix installed checking on Windows to avoid download on every time restart Zed. (#25)
Fixes #9

~maybe also #19~ nvm, reading more carefully I don't think it's related

The name of the language server executable for Windows is
`color-lsp.exe`; this should be used when checking for a pre-existing
download, or when checking if the server is installed already.

I also added an extra protection so that if the check for update fails
but there is already a binary installed, the status gets reset to `None`
and not failed/downloading.

---------

Co-authored-by: Jason Lee <huacnlee@gmail.com>